Enhanced ch3-08-lock example
The book says on page 61:

"Thanks to the acquire and release memory ordering, we know for sure that no two
threads can concurrently access DATA."

That creates an impression that no two threads can concurrently access LOCKED,
as actually it's the AtomicBool that is accessed with the acquire and release
memory ordering, not the DATA String.

This is further reassured by the Figure 3-4 on page 62, presenting the happens-before
relationships between atomic operations.

However this seems to be not the case in reality, and this simple modification
makes this easy to see that this is actually NOT the case, and more than one
thread can concurrently access LOCKED, and thus some threads fail to (re-)set
LOCKED to true, as running this example several times will clearly show that not
always we'll get a clean run and 100 exclamation signs.

I.e.:
ThreadId(51) Error - LOCKED is already set to true.
ThreadId(85) Error - LOCKED is already set to true.
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!, 98

Signed-off-by: Lev Veyde <lveyde@gmail.com>